requirements.txt
django~=4.2.1
python-dotenv~=1.0.0